ap-web/requirements.txt
# Exact pins (supply-chain hygiene): rebuild the image and re-test before
# bumping any of these. The flask, flask-cors, and requests pins close
# CVEs flagged by Aikido on 2026-05-01.
flask==3.1.3
flask-cors==6.0.0
gunicorn==25.3.0
psycopg2-binary==2.9.10
pyyaml==6.0.2
requests==2.33.0
websockets==13.1
